@@ -17,21 +17,21 @@
         }
     ],
     "require": {
-        "php": ">=5.5.9",
+        "php": ">=7.2",
         "doctrine/collections": "^1.3",
-        "symfony/form": "^3.4|^4.0",
-        "symfony/framework-bundle": "^3.4|^4.0",
-        "symfony/security": "^3.4|^4.0",
-        "symfony/twig-bundle": "^3.4|^4.0"
+        "symfony/form": "^5.0|^6.0",
+        "symfony/framework-bundle": "^5.0|^6.0",
+        "symfony/security-bundle": "^5.0|^6.0",
+        "symfony/twig-bundle": "^5.0|^6.0"
     },
     "require-dev": {
-        "symfony/validator": "^3.4|^4.0",
-        "symfony/translation": "^3.4|^4.0",
-        "symfony/yaml": "^3.4|^4.0",
-        "symfony/security-bundle": "^3.4|^4.0",
-        "symfony/templating": "^3.4|^4.0",
-        "symfony/browser-kit": "^3.4|^4.0",
-        "symfony/phpunit-bridge": "^4.2.8",
+        "symfony/validator": "^5.0|^6.0",
+        "symfony/translation": "^5.0|^6.0",
+        "symfony/yaml": "^5.0|^6.0",
+        "symfony/security-bundle": "^5.0|^6.0",
+        "symfony/templating": "^5.0|^6.0",
+        "symfony/browser-kit": "^5.0|^6.0",
+        "symfony/phpunit-bridge": "^6.1",
         "doctrine/orm": "^2.0"
     },
     "suggest": {
